bid handling done going for category then watchlist

+++ v2/2.2/commerce/auctions/models.py
@@ -21,7 +21,7 @@ class Bid(models.Model):
     bid_id=models.ForeignKey(Listing, on_delete=models.CASCADE, related_name="current_bidder_id")
     time = models.DateTimeField(auto_now_add=True, blank=True)
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name="user_bids")
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.IntegerField()
 
     def __str__(self):
         return f"{self.user} put a bid in for {self.price}"
